refactor slide
