+++ src/sass/components/_slider.scss
@@ -121,6 +121,10 @@
 @media only screen and (max-width: 600px) {
   .slider-container {
     height: 500px;
+
+    &.mobile {
+      height: 205px;
+    }
   }
 }
